refactor: update slimcat imports to use services
the previous commit added index.ts to electron/services. this makes slimcat imports look way cleaner and less ugly and hacky.

Authored by:
Rose (rose@rosely.me)
